chore(ngx-pass-code): rename transform-uppercase.directive to match selector
The file was named after a single feature (uppercase) while the
selector, class, and exported symbol are TransformInputValue/
transformInputValue. Rename the file and its spec so name, selector,
and class line up. Public API (TransformInputValueDirective from
ngx-pass-code) is unchanged.
