+++ libs/ngx-pass-code/src/lib/directives/transform-input-value.directive.ts
